+++ src/GitHub/commands.ts
@@ -1,7 +1,10 @@
 import * as rest from "@octokit/rest";
+import { TextDecoder } from "util";
+import { store } from "../FileSystem/store";
 import { RepoNode } from "../Tree/nodes";
-import { credentials } from "./../extension";
-import { TRepo, TGitHubTree, TGitHubUser, TGitHubBranch, TGitHubBranchList, TRepoContent } from './types';
+import { credentials, gitHubAuthenticatedUser, output } from "./../extension";
+import { COMMIT_MESSAGE } from "./constants";
+import { TRepo, TGitHubTree, TGitHubUser, TGitHubBranch, TGitHubBranchList, TRepoContent, TGitHubUpdateContent } from "./types";
 
 /**
  * Get the authenticated GitHub user
@@ -10,10 +13,11 @@ import { TRepo, TGitHubTree, TGitHubUser, TGitHubBranch, TGitHubBranchList, TRep
  * @async
  * @returns {Promise<TGitHubUser>}
  */
-export async function getGitHubUser(): Promise<TGitHubUser> {
+export async function getGitHubAuthenticatedUser(): Promise<TGitHubUser> {
     const octokit = new rest.Octokit({
         auth: await credentials.getAccessToken(),
     });
+
     const { data } = await octokit.users.getAuthenticated();
 
     return Promise.resolve(data);
@@ -26,16 +30,22 @@ export async function getGitHubUser(): Promise<TGitHubUser> {
  * @async
  * @returns {Promise<TRepo[]>}
  */
-export async function getGitHubReposForAuthenticatedUser(): Promise<TRepo[]> {
+export async function getGitHubReposForAuthenticatedUser(): Promise<TRepo[] | undefined> {
     const octokit = new rest.Octokit({
         auth: await credentials.getAccessToken(),
     });
 
-    const { data } = await octokit.repos.listForAuthenticatedUser({
-        type: "owner",
-    });
+    try {
+        const { data } = await octokit.repos.listForAuthenticatedUser({
+            type: "owner",
+        });
 
-    return Promise.resolve(data);
+        return Promise.resolve(data);
+    } catch (e: any) {
+        output.appendLine(`Could not get repositories for the authenticated user. ${e.message}`, output.messageType.error);
+    }
+
+    return Promise.reject(undefined);
 }
 
 /**
@@ -74,16 +84,43 @@ export async function getGitHubRepoContent(owner: string, repoName: string, path
  * @param {string} filePath The path to the file in the repository.
  * @returns {Promise<Uint8Array>}
  */
-export async function getRepoFile(repo: RepoNode, file: TRepoContent): Promise<Uint8Array> {
+export async function getRepoFileContent(repo: RepoNode, file: TRepoContent): Promise<Uint8Array> {
     let data: any;
     if (!file.content) {
         data = await getGitHubRepoContent(repo.owner, repo.name, file.path);
         file.content = data;
-    } else { 
+    } else {
         data = file.content;
     }
 
     return new Uint8Array(Buffer.from(data.content, "base64").toString("latin1").split("").map(charCodeAt));
+}
+
+export async function setRepoFileContent(repo: RepoNode, file: TRepoContent, content: Uint8Array): Promise<TGitHubUpdateContent> {
+    const fileContentString = (file.content = new TextDecoder().decode(content));
+    file.content = fileContentString;
+
+    const octokit = new rest.Octokit({
+        auth: await credentials.getAccessToken(),
+    });
+
+    try {
+        const { data } = await octokit.repos.createOrUpdateFileContents({
+            owner: repo.owner,
+            repo: repo.name,
+            path: file.path,
+            message: `${COMMIT_MESSAGE} ${file.path}`,
+            // message: `Repos: update file ${file.path}`,
+            content: Buffer.from(fileContentString).toString("base64"),
+            sha: file.sha,
+        });
+
+        return Promise.resolve(data);
+    } catch (e: any) {
+        output.logError(repo.repo, e);
+    }
+
+    return Promise.reject();
 }
 
 /**
@@ -106,19 +143,25 @@ function charCodeAt(c: string) {
  * @param {string} treeSHA
  * @returns {Promise<TGitHubTree>}
  */
-export async function getGitHubTree(repoOwner: string, repoName: string, treeSHA: string): Promise<TGitHubTree> {
+export async function getGitHubTree(repo: TRepo, treeSHA: string): Promise<TGitHubTree | undefined> {
     const octokit = new rest.Octokit({
         auth: await credentials.getAccessToken(),
     });
 
-    const { data } = await octokit.git.getTree({
-        owner: repoOwner,
-        repo: repoName,
-        tree_sha: treeSHA,
-        recursive: "true",
-    });
+    try {
+        const { data } = await octokit.git.getTree({
+            owner: repo.owner.login,
+            repo: repo.name,
+            tree_sha: treeSHA,
+            recursive: "true",
+        });
 
-    return Promise.resolve(data);
+        return Promise.resolve(data);
+    } catch (e: any) {
+        output.logError(repo, e);
+    }
+
+    return Promise.reject(undefined);
 }
 
 /**
@@ -130,42 +173,93 @@ export async function getGitHubTree(repoOwner: string, repoName: string, treeSHA
  * @param {string} repoName The name of the repo
  * @returns {Promise<TRepo>}
  */
-export async function getGitHubRepo(repoOwner: string, repoName: string): Promise<TRepo> {
+export async function getGitHubRepo(repo: TRepo, repoName: string): Promise<TRepo | undefined> {
     const octokit = new rest.Octokit({
         auth: await credentials.getAccessToken(),
     });
 
-    const { data } = await octokit.repos.get({
-        owner: repoOwner,
-        repo: repoName,
-    });
+    try {
+        const { data } = await octokit.repos.get({
+            owner: repo.owner.login,
+            repo: repoName,
+        });
 
-    return Promise.resolve(data);
+        return Promise.resolve(data);
+    } catch (e: any) {
+        output.logError(repo, e);
+    }
+
+    return Promise.reject(undefined);
 }
 
-export async function getGitHubBranch(repo: TRepo, branchName: string): Promise<TGitHubBranch> {
+export async function getGitHubBranch(repo: TRepo, branchName: string): Promise<TGitHubBranch | undefined> {
     const octokit = new rest.Octokit({
         auth: await credentials.getAccessToken(),
     });
 
-    const { data } = await octokit.repos.getBranch({
-        owner: repo.owner.login,
-        repo: repo.name,
-        branch: branchName,
-    });
+    try {
+        const { data } = await octokit.repos.getBranch({
+            owner: repo.owner.login,
+            repo: repo.name,
+            branch: branchName,
+        });
 
-    return Promise.resolve(data);
+        return Promise.resolve(data);
+    } catch (e: any) {
+        output.logError(repo, e);
+    }
+
+    return undefined;
 }
 
-export async function listGitHubBranches(repo: TRepo): Promise<TGitHubBranchList[]> {
+export async function listGitHubBranches(repo: TRepo): Promise<TGitHubBranchList[] | undefined> {
     const octokit = new rest.Octokit({
         auth: await credentials.getAccessToken(),
     });
 
-    const { data } = await octokit.repos.listBranches({
-        owner: repo.owner.login,
-        repo: repo.name,
+    try {
+        const { data } = await octokit.repos.listBranches({
+            owner: repo.owner.login,
+            repo: repo.name,
+        });
+
+        return Promise.resolve(data);
+    } catch (e: any) {
+        output.logError(repo, e);
+    }
+
+    return Promise.reject(undefined);
+}
+
+export async function createFolder(): Promise<void> {
+    throw new Error("Not implemented");
+}
+
+export async function createOrUpdateFile(repo: TRepo, path: string, content: string, sha?: string): Promise<void> {
+    const octokit = new rest.Octokit({
+        auth: await credentials.getAccessToken(),
     });
 
-    return Promise.resolve(data);
+    try {
+        const { data } = await octokit.repos.createOrUpdateFileContents({
+            owner: repo.owner.login,
+            repo: repo.name,
+            path,
+            message: "Created file",
+            // committer: {
+            //     name: gitHubAuthenticatedUser.name!,
+            //     email: gitHubAuthenticatedUser.email!,
+            // },
+            content,
+            sha: sha,
+        });
+
+        // todo: update the virtual file system
+
+        return Promise.resolve();
+    } catch (e: any) {
+        output.logError(repo, e);
+    }
+
+    return Promise.reject();
 }