+++ src/GitHub/constants.ts
@@ -1,1 +1,2 @@
-const DEFAULT_BRANCH = "master";
+export const DEFAULT_BRANCH = "master";
+export const COMMIT_MESSAGE = "Repos: update file";