+++ src/GitHub/types.ts
@@ -384,4 +384,67 @@ export type TGitHubBranch = {
     protection_url: string;
 };
 
-// export type TGitHubBranch = GetResponseDataTypeFromEndpointMethod<typeof octokit.repos.getBranch>;
+export type TGitHubUpdateContent = {
+    commit?:
+        | {
+              sha?: string | undefined;
+              node_id?: string | undefined;
+              url?: string | undefined;
+              html_url?: string | undefined;
+              author?:
+                  | {
+                        name?: string | undefined;
+                        email?: string | undefined;
+                        date?: string | undefined;
+                    }
+                  | undefined;
+              committer?:
+                  | {
+                        name?: string | undefined;
+                        email?: string | undefined;
+                        date?: string | undefined;
+                    }
+                  | undefined;
+              tree?:
+                  | {
+                        sha?: string | undefined;
+                        url?: string | undefined;
+                    }
+                  | undefined;
+              message?: string | undefined;
+              parents?:
+                  | {
+                        url?: string | undefined;
+                        html_url?: string | undefined;
+                        sha?: string | undefined;
+                    }[]
+                  | undefined;
+              verification?:
+                  | {
+                        verified?: boolean | undefined;
+                        reason?: string | undefined;
+                        signature?: string | null | undefined;
+                        payload?: string | null | undefined;
+                    }
+                  | undefined;
+          }
+        | undefined;
+    content: {
+        name?: string | undefined;
+        path?: string | undefined;
+        sha?: string | undefined;
+        size?: number | undefined;
+        url?: string | undefined;
+        html_url?: string | undefined;
+        git_url?: string | undefined;
+        download_url?: string | undefined;
+        type?: string | undefined;
+        _links?:
+            | {
+                  self?: string | undefined;
+                  git?: string | undefined;
+                  html?: string | undefined;
+              }
+            | undefined;
+    } | null;
+};