+++ src/extension.ts
@@ -4,14 +4,19 @@ import * as trace from "./tracing";
 import { commands, ExtensionContext, workspace, window } from "vscode";
 import { RepoProvider } from "./Tree/nodes";
 import { RepoFileSystemProvider, REPO_SCHEME } from "./FileSystem/fileSystem";
+import { getGitHubAuthenticatedUser } from "./GitHub/commands";
+import { TGitHubUser } from './GitHub/types';
 
 export let output: trace.Output;
 export const credentials = new Credentials();
+export let gitHubAuthenticatedUser: TGitHubUser;
 
 export async function activate(context: ExtensionContext) {
     if (config.get("EnableTracing")) {
         output = new trace.Output();
     }
+
+    gitHubAuthenticatedUser = await getGitHubAuthenticatedUser();
 
     output?.appendLine("Repos extension is now active!", output.messageType.info);
 
@@ -31,6 +36,24 @@ export async function activate(context: ExtensionContext) {
 
     context.subscriptions.push(
         commands.registerCommand("Repos.refreshTree", async () => {
+            repoProvider.refresh();
+        })
+    );
+
+    context.subscriptions.push(
+        commands.registerCommand("Repos.openRepository", async () => {
+            repoProvider.refresh();
+        })
+    );
+
+    context.subscriptions.push(
+        commands.registerCommand("Repos.addFolder", async () => {
+            repoProvider.refresh();
+        })
+    );
+
+    context.subscriptions.push(
+        commands.registerCommand("Repos.addFile", async () => {
             repoProvider.refresh();
         })
     );